Face Recognition/cloudguruamit/requirements.txt
streamlit
boto3
matplotlib
scikit-image
Pillow
langchain
